==> scripts/20_create_synthetic_data.ipynb <==
# References:

SDV is a public, source-available Python library for generating and evaluating synthetic data. You can download and use it under the Business Source License.

License: https://github.com/sdv-dev/SDV/blob/main/LICENSE

In this notebook, I've demonstrated a series of steps involved in generating synthetic transaction data using the sdv library in Python. Starting with loading the original transaction data into a DataFrame, I proceeded to create metadata using the SingleTableMetadata class from sdv. This metadata was then used to synthesize new data using the GaussianCopulaSynthesizer, incorporating constraints to maintain fixed combinations of certain columns. Quality evaluation and diagnostic assessment were conducted to compare the synthetic data with the original dataset, ensuring its fidelity. Additionally, measures were taken to generate consistent primary key generation like order IDs. Finally, the synthetic transaction data was saved to a CSV file for further analysis.

## Reading and Displaying Clean Olist Data

This cell imports the necessary library and reads the cleaned Olist data (created in the previous notebook step) from a CSV file into a pandas DataFrame. It then displays the first few rows of the DataFrame.

# Importing the pandas library as pd
import pandas as pd

# Reading the clean Olist data from a CSV file into a DataFrame
masterdf = pd.read_csv(
    '../data/clean_olist_data.csv'
)

# Displaying the first few rows of the DataFrame
masterdf.head()

# Checking the shape of the DataFrame
masterdf.shape

# Displaying information about the DataFrame
masterdf.info()

# Filtering the DataFrame to select rows where the 'order_id' column matches a specific value
masterdf[masterdf['order_id']=='ca3625898fbd48669d50701aba51cd5f']

## Generating Metadata for DataFrame with sdv

This cell utilizes the sdv library to automatically detect metadata from the provided DataFrame (`masterdf`). It then validates the detected metadata and converts it into a Python dictionary for further use.

from sdv.metadata import SingleTableMetadata

# Creating SingleTableMetadata object
metadata = SingleTableMetadata()

# Detecting metadata from the DataFrame. As an optional step we can save it to json for reproducibility.
metadata.detect_from_dataframe(masterdf)

# Validating the detected metadata
if metadata.validate() is not None:
    print("Problems with metadata auto-detection!")

# Converting metadata to a dictionary
python_dict = metadata.to_dict()

# Displaying the dictionary
python_dict

## Extracting Columns from Metadata

In this cell, we extract the names of all columns from the metadata obtained using the sdv library. The 'order_id' column is excluded from the list of all columns meant to be converted to categorical.

# Extracting all column names from metadata
all_cols = metadata.get_column_names()

# Identifying columns to be converted to categorical by excluding 'order_id'
all_cat_cols = list(set(all_cols) - set(['order_id']))

## Generating Synthetic Data with Gaussian Copula Synthesizer and Constraints

This cell uses the Gaussian Copula Synthesizer from the sdv library to generate synthetic data based on the provided metadata. Constraints are applied to maintain fixed combinations of certain columns, such as user and product identifiers. The resulting synthetic data is displayed to provide an overview.

from sdv.single_table import GaussianCopulaSynthesizer

# Updating metadata columns to categorical
metadata.update_columns(
    column_names=all_cat_cols,
    sdtype='categorical',
)

# Initializing GaussianCopulaSynthesizer with metadata and enforcing min-max values
synthesizer = GaussianCopulaSynthesizer(
    metadata,
    enforce_min_max_values=True,
)

# Defining user and product constraints
user_constraint = {
    'constraint_class': 'FixedCombinations',
    'constraint_parameters': {
        'column_names': ['user_id', 'customer_city']
    }
}
product_constraint = {
    'constraint_class': 'FixedCombinations',
    'constraint_parameters': {
        'column_names': ['product_id', 'product_category']
    }
}

# Adding constraints to the synthesizer
synthesizer.add_constraints(constraints=[
    user_constraint,
    product_constraint
])

# Fitting the synthesizer to the original data
synthesizer.fit(masterdf)

# Generating synthetic data
synthetic_data = synthesizer.sample(num_rows=10000000)

# Displaying the first few rows of synthetic data
synthetic_data.head()

## Loading and Validating Metadata from Python Dictionary

In this cell, the metadata for the DataFrame is loaded from a Python dictionary using the sdv library. The loaded metadata is then validated to ensure its integrity. If any issues are detected during the validation process, a corresponding message is printed.

from sdv.metadata import SingleTableMetadata

# Loading metadata from the Python dictionary
metadata = SingleTableMetadata.load_from_dict(python_dict)

# Validating the loaded metadata
if metadata.validate() is not None:
    print("Problems with resetting metadata detected!")

## Running Diagnostic Evaluation between Real and Synthetic Data

In this cell, a diagnostic evaluation is performed to compare the characteristics of the real data (`masterdf`) with the synthetic data (`synthetic_data`) generated using the Gaussian Copula Synthesizer. The evaluation is conducted using the provided metadata.

from sdv.evaluation.single_table import run_diagnostic

# Running diagnostic evaluation
diagnostic_report = run_diagnostic(
    real_data=masterdf,
    synthetic_data=synthetic_data,
    metadata=metadata
)

## Evaluating the Quality of Synthetic Data

This cell assesses the quality of synthetic data generated using the Gaussian Copula Synthesizer by comparing it with the real data (`masterdf`). The evaluation is performed using various quality metrics and the provided metadata.

from sdv.evaluation.single_table import evaluate_quality

# Evaluating the quality of synthetic data
quality_report = evaluate_quality(
    real_data=masterdf,
    synthetic_data=synthetic_data,
    metadata=metadata
)

quality_report.get_details(property_name='Column Shapes')

## Creating primary key Order IDs in Synthetic Data

This cell anonymizes the order IDs in the synthetic data by combining the 'timestamp' and 'user_id' columns, hashing the combined values using SHA256, and storing the result as the new 'order_id' column. Any duplicate rows in the synthetic data are then removed to ensure data integrity.

import hashlib

# Combining 'timestamp' and 'user_id' columns and hashing the combined values to create 'order_id'
synthetic_data['combined_values'] = synthetic_data['timestamp'].astype(str) + '-' + synthetic_data['user_id']
synthetic_data['order_id'] = synthetic_data['combined_values'].apply(lambda x: hashlib.sha256(x.encode()).hexdigest()[:32])

# Dropping the 'combined_values' column
synthetic_data = synthetic_data.drop(columns=['combined_values'])

# Dropping duplicate rows from the synthetic data
synthetic_data.drop_duplicates(inplace=True)

synthetic_data[['timestamp', 'user_id', 'customer_city']].value_counts().reset_index().loc[0]

synthetic_data[synthetic_data["user_id"]=="7c396fd4830fd04220f754e42b4e5bff"]

len(synthetic_data)

## Saving Synthetic Transaction Data to CSV

This cell saves the synthetic transaction data stored in the DataFrame `synthetic_data` to a CSV file named 'all_transaction_data.csv' in the '../data/' directory. The index column is excluded from the CSV file.

synthetic_data.to_csv('../data/all_transaction_data.csv', index=False)

By leveraging sdv, I've generated synthetic dataset that preserve key statistical properties and relationships present in the original data. However, it's essential to carefully evaluate the quality and integrity of synthetic data before use, ensuring it accurately reflects the underlying patterns of the real-world data. In subsequent notebooks, this dataset will serve as the foundation for constructing rank-based recommendation models. Additionally, it will be harnessed to furnish data corresponding to user queries in later stages.
