==> scripts/50_message_data_preperation.ipynb <==
# Data Analysis and Preperation

This script conducts data analysis tasks on a CRM dataset using Python's Pandas and NumPy libraries. Initially, the data is loaded from a CSV file, and optional preprocessing steps are applied, such as removing backticks from a specific column. Subsequently, the dataset is split into training, validation, and test sets for machine learning purposes. A message template is defined to generate structured conversation data, which includes system prompts, user questions, and assistant responses. Finally, the datasets are shuffled and saved into JSON files for further processing.

# Importing necessary libraries
import pandas as pd
import numpy as np

np.random.seed(42)  # for reproducibility

# Reading data
data = pd.read_csv('../data/CRM_data_no_comments.csv')
len(data)

# Removing backticks (`) from the 'context' column in the DataFrame [Optional]
data['context'] = data['context'].apply(lambda x: x.replace('`', ''))
data.head()

# Splitting data into train, validation and test sets

shuffled_indices = np.random.permutation(len(data)) # Shuffle the indices

# Define the proportions for each set
train_prop = 0.8
val_prop = 0.1
test_prop = 0.1

# Calculate the sizes of each set
num_train = int(len(data) * train_prop)
num_val = int(len(data) * val_prop)
num_test = len(data) - num_train - num_val

# Split the shuffled indices into train, validation, and test sets
train_indices = shuffled_indices[:num_train]
val_indices = shuffled_indices[num_train:num_train + num_val]
test_indices = shuffled_indices[num_train + num_val:]

# Create the subsets using the indices
train_df = data.iloc[train_indices]
val_df = data.iloc[val_indices]
test_df = data.iloc[test_indices]

# Define a message template
system_message = system_message = """
You are a Python function generator. Users will ask you questions in English, 
and you will produce a Python function as answer to the question based on the provided CONTEXT.

CONTEXT:
{context}
"""

def create_messages(sample):
    """
    Create a message structure from a sample containing context, question, and solution.

    Args:
        sample (pd.Series): A pandas Series containing 'context', 'question', and 'solution_no_comments' fields.

    Returns:
        dict: A dictionary containing messages structured for a conversation. Each message includes a role
              (system, user, or assistant) and content (context, question, or solution).
    """
    return {
        "messages": [
            {"role": "system", "content": system_message.format(context=sample["context"].replace('`', ''))},
            {"role": "user", "content": sample["question"].replace('`', '')},
            {"role": "assistant", "content": sample["solution_no_comments"].replace('`', '')}
        ]
    }

# Apply create_messages function to create datasets
train_dataset = train_df.apply(create_messages, axis=1)
val_dataset = val_df.apply(create_messages, axis=1)
test_dataset = test_df.apply(create_messages, axis=1)

# Shuffle the datasets
random_train_dataset = train_dataset.sample(frac=1).reset_index(drop=True)
random_val_dataset = val_dataset.sample(frac=1).reset_index(drop=True)
random_test_dataset = test_dataset.sample(frac=1).reset_index(drop=True)

# Save datasets to JSON files
random_train_dataset.to_json("../data/train_CRM_data.json", orient="records")
random_val_dataset.to_json("../data/val_CRM_data.json", orient="records")
random_test_dataset.to_json("../data/test_CRM_data.json", orient="records")

These structured message datasets provides the foundation for training of text generation AI models to follow in the subsequent steps.
